requirements.txt
anthropic>=0.40.0
langgraph>=0.2.0
pyyaml>=6.0
